fix: log error when ssr render failed
